Capstone Project 2
This Capstone project is a report generated in Power BI as part of an academic study and with a guide as shared in the pdf file. 

This visualization is to communicate the outcome of our analysis to our investors who are both technical and non-technical stakeholders of the company on which field to make investments based on the global sales of the sampled video games. The dataset is shared in Excel format as vgsales-vgsales.xlsx

The documentation attached is the word format of our analysis and insights from the data story.